Refactors out structured parsing logic to a separate module.
